feat: creating wrapper to home page
